clase

+++ Programacion/Java/Tkinter/src/Calculadora.java
@@ -4,7 +4,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
 public class Calculadora {
-    static double botonPorDefecto = 0;
+    static int botonPorDefecto = 0;
     static boolean botonPresionado = false;
     public static void main(String[] args) {
         JFrame frame = new JFrame("Calculadora");
